webempresa/webempresa/requirements.txt
Django==2.0
django-ckeditor==5.6.1
django-extensions==2.1.2
django-js-asset==1.1.0
django-tinymce4-widget==4.0.0
Pillow==5.2.0
pytz==2018.5
six==1.11.0
